Systematize one-off data conversion scripts
